Abstract pruning + implement txn accounts pruning
