+++ src/test_suite/fuzz_context.py
@@ -4,7 +4,9 @@ import test_suite.invoke_pb2 as invoke_pb
 import test_suite.elf_pb2 as elf_pb
 import test_suite.vm_pb2 as vm_pb
 import test_suite.txn.codec_utils as txn_codec
+import test_suite.txn.prune_utils as txn_prune
 import test_suite.instr.codec_utils as instr_codec
+import test_suite.instr.prune_utils as instr_prune
 import test_suite.syscall.codec_utils as syscall_codec
 
 
@@ -16,6 +18,7 @@ ElfHarness = HarnessCtx(
 InstrHarness = HarnessCtx(
     fuzz_fn_name="sol_compat_instr_execute_v1",
     fixture_desc=invoke_pb.InstrFixture.DESCRIPTOR,
+    prune_effects_fn=instr_prune.prune_execution_result,
     context_human_encode_fn=instr_codec.encode_input,
     context_human_decode_fn=instr_codec.decode_input,
     effects_human_encode_fn=instr_codec.encode_output,
@@ -36,6 +39,7 @@ ValidateVM = HarnessCtx(
 TxnHarness = HarnessCtx(
     fuzz_fn_name="sol_compat_txn_execute_v1",
     fixture_desc=txn_pb.TxnFixture.DESCRIPTOR,
+    prune_effects_fn=txn_prune.prune_execution_result,
     context_human_encode_fn=txn_codec.encode_input,
     context_human_decode_fn=txn_codec.decode_input,
     effects_human_encode_fn=txn_codec.encode_output,
