+++ src/test_suite/fuzz_interface.py
@@ -19,10 +19,17 @@ The following defines the interface:
     - input: The fuzz target Context
     - output: The fuzz target Effects
 - diff_effect_fn: A function that compares two effects messages for equality
+- prune_effects_fn: A function that prunes effects to remove extra fields (e.g. remove accounts that weren't actually modified)
 - human encode/decode functions for the context and effects messages to
   convert the messages to/from human-readable format (in-place).
   Both context and effects messages can have their own encode/decode functions.
 """
+
+
+def generic_effects_prune(
+    ctx: ContextType, effects: dict[str, EffectsType]
+) -> dict[str, EffectsType]:
+    return effects
 
 
 def generic_effects_diff(a: EffectsType, b: EffectsType) -> bool:
@@ -44,6 +51,9 @@ class HarnessCtx:
     result_field_name: str | None = "result"
     ignore_fields_for_consensus: list[str] = field(default_factory=list)
     diff_effect_fn: Callable[[EffectsType, EffectsType], bool] = generic_effects_diff
+    prune_effects_fn: Callable[
+        [ContextType, dict[str, EffectsType]], dict[str, EffectsType]
+    ] = generic_effects_prune
     context_human_encode_fn: Callable[[ContextType], None] = generic_human_encode
     context_human_decode_fn: Callable[[ContextType], None] = generic_human_decode
     effects_human_encode_fn: Callable[[EffectsType], None] = generic_human_encode
